+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Queues/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 5.0.0 (2021-10-21)
 - General availability of Microsoft.Azure.WebJobs.Extensions.Storage.Queues 5.0.0.
-- Change `QueueProcessor.MessageAddedToPoisonQueue` to async event.
+- Change `QueueProcessor.MessageAddedToPoisonQueue` to async event and rename to `QueueProcessor.MessageAddedToPoisonQueueAsync`.
 - QueuesOptions.MaxPollingInterval other than default is now honored in "Development" environment.
 - Adding Dynamic Concurrency support.